refactor:  Fix Router의 Component화
1. 왜? : App.js파일에 import가 과하게 많다고 판단했다.
2. 어떻게? : Router 관련 import한 것이 많으므로 Router를 Component화 했다.
